refactor: Pull PackageIdSpec into schema
