+++ src/cargo/core/mod.rs
@@ -4,7 +4,7 @@ pub use self::manifest::{EitherManifest, VirtualManifest};
 pub use self::manifest::{Manifest, Target, TargetKind};
 pub use self::package::{Package, PackageSet};
 pub use self::package_id::PackageId;
-pub use self::package_id_spec::{PackageIdSpec, PackageIdSpecQuery};
+pub use self::package_id_spec::PackageIdSpecQuery;
 pub use self::registry::Registry;
 pub use self::resolver::{Resolve, ResolveVersion};
 pub use self::shell::{Shell, Verbosity};
@@ -14,7 +14,7 @@ pub use self::workspace::{
     find_workspace_root, resolve_relative_path, MaybePackage, Workspace, WorkspaceConfig,
     WorkspaceRootConfig,
 };
-pub use crate::util_schemas::core::{GitReference, SourceKind};
+pub use crate::util_schemas::core::{GitReference, PackageIdSpec, SourceKind};
 
 pub mod compiler;
 pub mod dependency;
