+++ src/cargo/util_schemas/core/mod.rs
@@ -1,4 +1,6 @@
+mod package_id_spec;
 mod source_kind;
 
+pub use package_id_spec::PackageIdSpec;
 pub use source_kind::GitReference;
 pub use source_kind::SourceKind;
